+++ packages/react-ui/storybook/stories/molecules/Breadcrumb.stories.js
@@ -19,7 +19,7 @@ const options = {
       url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1534%3A26153'
     },
     status: {
-      type: 'needUpdate'
+      type: 'needsUpdate'
     }
   }
 };