+++ packages/react-ui/storybook/stories/organisms/Dialog.stories.js
@@ -54,7 +54,7 @@ const options = {
       url: 'https://www.figma.com/file/nmaoLeo69xBJCHm9nc6lEV/CARTO-Components-1.0?node-id=1534%3A28896'
     },
     status: {
-      type: 'needUpdate'
+      type: 'needsUpdate'
     }
   }
 };