+++ packages/react-ui/storybook/stories/widgets/PieWidget.stories.js
@@ -12,7 +12,7 @@ import * as cartoSlice from '../../../../react-redux/src/slices/cartoSlice';
 import { AggregationTypes } from '../../../../react-core/src';
 import PieWidget from '../../../../react-widgets/src/widgets/PieWidget';
 import { mockAppStoreConfiguration } from './utils';
-import { buildReactPropsAsString } from '../../utils';
+import { buildReactPropsAsString } from '../../utils/utils';
 
 const store = mockAppStoreConfiguration();