+++ packages/react-ui/storybook/stories/widgetsUI/NoDataAlert.stories.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import NoDataAlert from '../../../src/widgets/NoDataAlert';
-import { buildReactPropsAsString } from '../../utils';
+import { buildReactPropsAsString } from '../../utils/utils';
 
 const options = {
   title: 'Organisms/Widgets/NoDataAlert',